Update ocropus-rtrain
Currently there is a little bug in making a new codec. The argument -c needs a list of input text files (nargs='+/*' missing). Or else, -c could be used to store an action and then use args.files to build a new codec. There are a lot of confusion among new users on how to train ocropus for new languages, and I found this bug. I've tested this with some languages and found the current modification working.