Explicitly encode output as UTF-8
I think I've covered all the cases we need to, but there could be more.

This is needed as if a locale isn't set to UTF-8, an error will result
of this form:

  UnicodeEncodeError: 'ascii' codec can't encode character u'\u0113' in position 60: ordinal not in range(128)

While the ideal solution is for the user to set their locale to UTF-8,
it is better that we print debug output which may not be displayed
correctly than that we output a fatal (and non-obvious) error,
potentially some time into processing.

This also fixes some cases of implicitly combining str and *obj together
when printing debug output, which fails with some Python versions, by
explicitly using str.join(obj).