Add NFC and NFD normalization options (keeping NFKC as the default)
While NFKC is kept as the default, this gives the option to use NFC
and NFD normalization options. These can't be used directly, but
allow a model that has been trained with an alternative
normalization to be loaded and used. Without this patch, such a
model will throw an error when unpickling such a model.

Such a model can be built using different normalization= parameters,
as for example in PR#257.